requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2024.7.4
charset-normalizer==3.2.0
discord.py==2.3.1
frozenlist==1.4.0
fuzzywuzzy==0.18.0
greenlet==2.0.2
idna==3.4
multidict==6.0.4
pydantic==1.10.12
requests==2.32.3
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.8
typing_extensions==4.7.1
urllib3==2.2.2
yarl==1.9.2
